* moved testing classes to own folder * fixed doc
