+++ slick/src/test/scala/za/co/absa/db/fadb/testng/classes/ActorSlickConverter.scala
